+++ model/schema_insert_test.go
@@ -1,0 +1,10 @@
+package model
+
+import "testing"
+
+func TestInsert(t *testing.T) {
+	// TODO: Schema.Insert
+	// TODO: Schema.InsertAt
+	// TODO: Schema.InsertTx
+	// TODO: Schema.InsertTxAt
+}